feat: can narrow pre blocks
User can `C-c C-c` in a pre blocks to open a new buffer with the
contained text and the adequate major mode. `C-c C-c` again to go back
to the Gemtext buffer.
Known issue: there must be at least one empty line in the pre block,
otherwise the command will not execute properly.

Kyle Anders <kyle@posteo.net> is the main author of this contribution.